bug修改

+++ app/src/main/java/com/slash/youth/ui/viewmodel/MyPublishServiceModel.java
@@ -967,7 +967,7 @@ public class MyPublishServiceModel extends BaseObservable {
                     setInstalmentVisibility(View.VISIBLE);
                     String instalmentRatioStr = "";
                     for (int i = 0; i < instalmentInfoList.size(); i++) {
-                        String ratio = instalmentInfoList.get(i).percent * 100 + "";
+                        String ratio = (int) (instalmentInfoList.get(i).percent * 100) + "";
                         if (TextUtils.isEmpty(ratio)) {
                             continue;
                         }
